Reorganized `particpants` slightly
