+++ server/services/participants/index.ts
@@ -1,8 +1,8 @@
 export * from './acknowledgement';
 export * from './change-history';
 export * from './confirmation';
-export * from './participant-single';
-export * from './participants-multi';
+export * from './participant-entries';
+export * from './batch';
 export * from './site-participants';
 export * from './status';
 export * from './user-participants';
