+++ server/services/participants/participant-entries.ts
@@ -1,6 +1,10 @@
-import readXlsxFile from 'node-xlsx';
-import { createRows, verifyHeaders } from '../../utils';
+import dayjs from 'dayjs';
 import { dbClient, collections } from '../../db';
+import { postHireStatuses } from '../../validation';
+import { getAssignCohort } from '../cohorts';
+import { createPostHireStatus, getPostHireStatusesForParticipant } from '../post-hire-flow';
+import logger from '../../logger';
+import { ParticipantsFinder } from '../participants-helper';
 import type {
   EmailAddressFilter,
   IsIndigenousFilter,
@@ -8,10 +12,162 @@ import type {
   Pagination,
   PostalCodeFsaFilter,
 } from '../participants-helper';
-import { ParticipantsFinder } from '../participants-helper';
-import { getPostHireStatusesForParticipant } from '../post-hire-flow';
-import { validate, ParticipantBatchSchema, isBooleanValue } from '../../validation';
-import { HcapUserInfo } from '../../keycloak';
+import type { HcapUserInfo } from '../../keycloak';
+
+export const makeParticipant = async (participantData) => {
+  const res = await dbClient.db.saveDoc(collections.PARTICIPANTS, participantData);
+  return res;
+};
+
+export const getParticipantByID = async (id) => {
+  const participant = await dbClient.db[collections.PARTICIPANTS].findDoc({
+    id,
+  });
+  return participant;
+};
+
+export const getParticipantByIdWithStatus = async ({ id, userId }) =>
+  dbClient.db[collections.PARTICIPANTS]
+    .join({
+      currentStatuses: {
+        type: 'LEFT OUTER',
+        relation: collections.PARTICIPANTS_STATUS,
+        on: {
+          participant_id: 'id',
+          current: true,
+        },
+      },
+      user: {
+        type: 'INNER',
+        relation: collections.USER_PARTICIPANT_MAP,
+        decomposeTo: 'object',
+        on: {
+          participant_id: 'id',
+          user_id: userId,
+        },
+      },
+    })
+    .find({ id, 'user.user_id': userId });
+
+export const deleteParticipant = async ({ email }) => {
+  await dbClient.db.withTransaction(async (tnx) => {
+    // Delete entry from participant-user-map
+    await tnx.query(
+      `
+        DELETE FROM ${collections.USER_PARTICIPANT_MAP} 
+        WHERE participant_id IN
+        ( SELECT id FROM ${collections.PARTICIPANTS} 
+          WHERE LOWER(body->>'emailAddress') = LOWER($1)
+        );`,
+      [email]
+    );
+    // Delete actual entry
+    await tnx[collections.PARTICIPANTS].destroy({
+      'body.emailAddress': email,
+    });
+  });
+};
+
+export const updateParticipant = async (participantInfo) => {
+  try {
+    // The below reduce function unpacks the most recent changes in the history
+    // and builds them into an object to be used for the update request
+    const changes = participantInfo.history[0].changes.reduce(
+      (acc, change) => {
+        const { field, to } = change;
+        return { ...acc, [field]: to };
+      },
+      { history: participantInfo.history || [], userUpdatedAt: new Date().toJSON() }
+    );
+    if (changes.interested === 'withdrawn') {
+      const cohort = await getAssignCohort({ participantId: participantInfo.id });
+      // Get All existing status
+      const statuses = await getPostHireStatusesForParticipant({
+        participantId: participantInfo.id,
+      });
+      const graduationStatuses = statuses.filter(
+        (item) =>
+          item.status === postHireStatuses.postSecondaryEducationCompleted ||
+          item.status === postHireStatuses.cohortUnsuccessful
+      );
+      // ensure that a participant has a cohort before adding post hire status
+      if (cohort && cohort.length > 0 && graduationStatuses.length === 0) {
+        await createPostHireStatus({
+          participantId: participantInfo.id,
+          status: postHireStatuses.cohortUnsuccessful,
+          data: {
+            unsuccessfulCohortDate: dayjs().format('YYYY/MM/DD'),
+          },
+        });
+      }
+    }
+    const participant = await dbClient.db[collections.PARTICIPANTS].updateDoc(
+      {
+        id: participantInfo.id,
+      },
+      changes
+    );
+
+    return participant;
+  } catch (error) {
+    logger.error(`updateParticipant: fail to update participant: ${error}`);
+    throw error;
+  }
+};
+
+export const setParticipantLastUpdated = async (id) => {
+  // Find participants
+  let [participant] = await getParticipantByID(id);
+  // Don't change status if participant is withdrawn
+  if (participant.interested !== 'withdrawn') {
+    // Only change history if the interested column isn't yes
+    if (participant.interested !== 'yes') {
+      if (participant.history) {
+        participant.history.push({
+          to: 'yes',
+          from: participant.interested,
+          field: 'interested',
+          timestamp: new Date(),
+        });
+      } else {
+        participant.history = [
+          {
+            to: 'yes',
+            from: participant.interested,
+            field: 'interested',
+            timestamp: new Date(),
+          },
+        ];
+      }
+    }
+    participant = await dbClient.db[collections.PARTICIPANTS].updateDoc(
+      {
+        id,
+      },
+      {
+        interested: 'yes',
+        history: participant.history,
+        userUpdatedAt: new Date().toJSON(),
+      }
+    );
+  }
+};
+
+export const withdrawParticipant = async (participantInfo) => {
+  const participant = { ...participantInfo };
+  const newHistory = {
+    timestamp: new Date(),
+    changes: [],
+  };
+  newHistory.changes.push({
+    field: 'interested',
+    from: participant.interested || 'yes',
+    to: 'withdrawn',
+  });
+  participant.history = participant.history ? [newHistory, ...participant.history] : [newHistory];
+  // eslint-disable-next-line no-use-before-define
+  return updateParticipant(participant);
+};
 
 export const getParticipants = async (
   user?: HcapUserInfo,
@@ -209,80 +365,4 @@ export const getParticipants = async (
     }),
     ...(pagination && { pagination: paginationData }),
   };
-};
-
-export const parseAndSaveParticipants = async (fileBuffer) => {
-  const columnMap = {
-    ClientID: 'maximusId',
-    Surname: 'lastName',
-    Name: 'firstName',
-    PostalCode: 'postalCode',
-    'Post Code FSA': 'postalCodeFsa',
-    Phone: 'phoneNumber',
-    Email: 'emailAddress',
-    'EOI - FHA': 'fraser',
-    'EOI - IHA': 'interior',
-    'EOI - NHA': 'northern',
-    'EOI - VCHA': 'vancouverCoastal',
-    'EOI - VIHA': 'vancouverIsland',
-    'CB1: Still Interested': 'interested',
-    'CB8: Non-HCAP Opportunities': 'nonHCAP',
-    'CB13: CRC Clear': 'crcClear',
-  };
-
-  const objectMap = (row) => {
-    const object = { ...row };
-
-    const preferredLocation = [];
-
-    if (row.fraser === 1 || isBooleanValue(row.fraser)) preferredLocation.push('Fraser');
-    if (row.interior === 1 || isBooleanValue(row.interior)) preferredLocation.push('Interior');
-    if (row.northern === 1 || isBooleanValue(row.northern)) preferredLocation.push('Northern');
-    if (row.vancouverCoastal === 1 || isBooleanValue(row.vancouverCoastal))
-      preferredLocation.push('Vancouver Coastal');
-    if (row.vancouverIsland === 1 || isBooleanValue(row.vancouverIsland))
-      preferredLocation.push('Vancouver Island');
-
-    object.preferredLocation = preferredLocation.join(';');
-    object.callbackStatus = false;
-    object.userUpdatedAt = new Date().toJSON();
-
-    delete object.fraser;
-    delete object.interior;
-    delete object.northern;
-    delete object.vancouverCoastal;
-    delete object.vancouverIsland;
-
-    return object;
-  };
-
-  const xlsx = readXlsxFile.parse(fileBuffer, { raw: true });
-  verifyHeaders(xlsx[0].data, columnMap);
-  let rows = createRows(xlsx[0].data, columnMap);
-  await validate(ParticipantBatchSchema, rows);
-  const lowercaseMixed = (v) => (typeof v === 'string' ? v.toLowerCase() : v);
-  rows = rows.map((row) => ({
-    ...row,
-    interested: lowercaseMixed(row.interested),
-  }));
-  const response = [];
-  const promises = rows.map((row) => dbClient.db.saveDoc(collections.PARTICIPANTS, objectMap(row)));
-  const results = await Promise.allSettled(promises);
-
-  results.forEach((result, index) => {
-    const id = rows[index].maximusId;
-    switch (result.status) {
-      case 'fulfilled':
-        // Update coordinates for all fulfilled promises
-        response.push({ id, status: 'Success' });
-        break;
-      default:
-        if (result.reason.code === '23505') {
-          response.push({ id, status: 'Duplicate' });
-        } else {
-          response.push({ id, status: 'Error', message: result.reason });
-        }
-    }
-  });
-  return response;
 };
